Move Runners hook into results hooks
