+++ src/pages/Results/shared/hooks.ts
@@ -1,6 +1,6 @@
 import {useCallback, useEffect, useState} from "react";
-import {getClassesInStage, getStageDetail} from "../services/EventService.ts";
-import {ClassModel} from "../../../shared/EntityTypes.ts";
+import {getClassesInStage, getRunnersInStage, getStageDetail} from "../services/EventService.ts";
+import {ClassModel, RunnerModel} from "../../../shared/EntityTypes.ts";
 import {useLocation, useParams, useSearchParams} from "react-router-dom";
 import {useAuth} from "../../../shared/hooks.ts";
 
@@ -122,4 +122,41 @@ export function useEventInfo() {
     ,[])
 
   return {eventId:eventId as string,eventName:eventName,stageId:stageId as string,stageName:stageName,stageTypeId:stageTypeId}
+}
+
+/**
+ * Get the runner list from the server associated with the event, stage and class. When the class
+ * is provided as a state, the list will be updated automatically if any of the three params gets
+ * updated.
+ * @param event_id
+ * @param stage_id
+ * @param activeClass
+ * @returns A list [runnerList,isLoading]
+ * @returns runnersList:RunnerModel[] state with the desired runners
+ * @returns isLoading:boolean state to know when result are being fetch
+ */
+export function useRunners(event_id:string,stage_id:string,activeClass:ClassModel|null):[RunnerModel[],boolean] {
+  const [isLoading,setIsLoading] = useState(true);
+  const [runnerList, setRunnerList] = useState<RunnerModel[]>([])
+
+  //fetch from back-end
+  useEffect(() => {
+    if (activeClass){
+      getRunnersInStage(event_id,stage_id,activeClass.id).then((response)=>{
+        setRunnerList(response.data)
+        setIsLoading(false)
+      },(error)=>{
+        console.log(error);
+        setIsLoading(false)
+      })
+    } else {
+      setRunnerList([])
+      setIsLoading(false);
+    }
+    return () => {
+      setIsLoading(true)
+    }
+  },[event_id,stage_id,activeClass])
+
+  return [runnerList,isLoading]
 }
